+++ src/version1/AccountsNullClientV1.ts
@@ -1,8 +1,8 @@
-import { IReferences } from 'pip-services-commons-node';
-import { Descriptor } from 'pip-services-commons-node';
-import { FilterParams } from 'pip-services-commons-node';
-import { PagingParams} from 'pip-services-commons-node';
-import { DataPage } from 'pip-services-commons-node';
+import { IReferences } from 'pip-services3-commons-node';
+import { Descriptor } from 'pip-services3-commons-node';
+import { FilterParams } from 'pip-services3-commons-node';
+import { PagingParams} from 'pip-services3-commons-node';
+import { DataPage } from 'pip-services3-commons-node';
 
 import { IAccountsClientV1 } from './IAccountsClientV1';
 import { AccountV1 } from './AccountV1';