+++ test/version1/AccountsCommandableGrpcClientV1.test.ts
@@ -8,12 +8,20 @@ import { ConsoleLogger } from 'pip-services3-components-node';
 
 import { AccountsMemoryPersistence } from 'pip-services-accounts-node';
 import { AccountsController } from 'pip-services-accounts-node';
+import { AccountsCommandableGrpcServiceV1 } from 'pip-services-accounts-node';
 import { IAccountsClientV1 } from '../../src/version1/IAccountsClientV1';
-import { AccountsDirectClientV1 } from '../../src/version1/AccountsDirectClientV1';
+import { AccountsCommandableGrpcClientV1 } from '../../src/version1/AccountsCommandableGrpcClientV1';
 import { AccountsClientFixtureV1 } from './AccountsClientFixtureV1';
 
-suite('AccountsDirectClientV1', ()=> {
-    let client: AccountsDirectClientV1;
+var httpConfig = ConfigParams.fromTuples(
+    "connection.protocol", "http",
+    "connection.host", "localhost",
+    "connection.port", 3000
+);
+
+suite('AccountsCommandableGrpcClientV1', ()=> {
+    let service: AccountsCommandableGrpcServiceV1;
+    let client: AccountsCommandableGrpcClientV1;
     let fixture: AccountsClientFixtureV1;
 
     suiteSetup((done) => {
@@ -21,23 +29,32 @@ suite('AccountsDirectClientV1', ()=> {
         let persistence = new AccountsMemoryPersistence();
         let controller = new AccountsController();
 
+        service = new AccountsCommandableGrpcServiceV1();
+        service.configure(httpConfig);
+
         let references: References = References.fromTuples(
             new Descriptor('pip-services', 'logger', 'console', 'default', '1.0'), logger,
             new Descriptor('pip-services-accounts', 'persistence', 'memory', 'default', '1.0'), persistence,
             new Descriptor('pip-services-accounts', 'controller', 'default', 'default', '1.0'), controller,
+            new Descriptor('pip-services-accounts', 'service', 'commandable-grpc', 'default', '1.0'), service
         );
         controller.setReferences(references);
+        service.setReferences(references);
 
-        client = new AccountsDirectClientV1();
+        client = new AccountsCommandableGrpcClientV1();
         client.setReferences(references);
+        client.configure(httpConfig);
 
         fixture = new AccountsClientFixtureV1(client);
 
-        client.open(null, done);
+        service.open(null, (err) => {
+            client.open(null, done);
+        });
     });
     
     suiteTeardown((done) => {
-        client.close(null, done);
+        client.close(null);
+        service.close(null, done);
     });
 
     test('CRUD Operations', (done) => {